+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.17.3] - 2021-04-09
+### Added
+- GS4 status query handler
+- Default values to all of the configuration options
+
+### Changed
+- Fixed server's address never getting instantiated
+- 
+
 ## [0.17.2] - 2021-04-08
 ### Changed
 - Internal vars in `KryptonServer` have been made `private set`, to avoid them being updated outside of